pyproject.toml
[project]
name = "hudson_platecrane_module"
version = "2.0.0"
description = "MADSci Node Module and interface for the Hudson Platecrane Robotic Arm"
authors = [
    {name = "Ryan D. Lewis", email="ryan.lewis@anl.gov"},
    {name = "Rafael Vescovi", email="ravescovi@anl.gov"},
    {name = "Doga Ozgulbas", email="dozgulbas@anl.gov"},
    {name = "Abe Stroka", email="astroka@anl.gov"},
    {name = "Kyle Hippe", email = "khippe@anl.gov"},
    {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
]
dependencies = [
    "pyserial",
    "madsci.node_module>=0.3.0",
]
requires-python = ">=3.9.1"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
homepage = "https://github.com/AD-SDL/hudson_platecrane_module"

######################
# Build Info + Tools #
######################
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.11.7",
    "pytest",
    "ipykernel>=6.29.5",
]
